refact(pkg): move apiutil to strategicmerge pkg
