+++ pkg/kubernetes/strategicmerge/v1alpha1/strategicmerge.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package apiutil
+package v1alpha1
 
 import (
 	"encoding/json"
